Normative: Prevent indefinite loops in NormalizedTimeDurationToDays
It's possible to make at least the second loop continue indefinitely with
a contrived calendar and time zone.

DRAFT: Still to be determined if this precludes any non-contrived use
cases. If so, we will keep the loops, but still put an upper limit on the
number of iterations.

Includes a few more tests in the NYSE time zone cookbook example to make
sure that a time zone transition of >24h continues to work.