Editorial: Update ECMA-402 original text to latest snapshot
Brings in the upstream ECMA-402 text as of commit tc39/ecma402@537afda7,
and adapts existing code on top of it.

Notably, ECMA-402 has refactored so that InitializeDateTimeFormat is
replaced by CreateDateTimeFormat, and ToDateTimeOptions is removed. This
refactor changed the observable order of accessing the options passed to
the Intl.DateTimeFormat constructors, in a way that was incompatible with
changes that Temporal made to the same order of accessing. This commit
reorders the new Temporal parts of CreateDateTimeFormat so that they no
longer change the observable order of operations.

Also corrects some ECMA-402 "rebase errors" that inadvertently made
normative changes in past updates.

Closes: #1932
Closes: #2363
Closes: #2473